Modify the data type of media_geometry3d from double to float to avoid inconsistent accuracy.

+++ media/media_geometry3d.cpp
@@ -5,7 +5,7 @@
 #include "media_geometry3d.hpp"
 //using namespace std;
 
-int dcmp(double x) {
+int dcmp(float x) {
 	if (fabs(x) < EPS) 
 		return 0;
 	else
@@ -20,32 +20,32 @@ Vector3 operator - (const Vector3 &A, const Vector3 &B) {
 	return Vector3(A.x-B.x, A.y-B.y, A.z-B.z);
 }
 
-Vector3 operator * (const Vector3 &A, double p) {
+Vector3 operator * (const Vector3 &A, float p) {
 	return Vector3(A.x*p, A.y*p, A.z*p);
 }
 
-Vector3 operator / (const Vector3 &A, double p) {
+Vector3 operator / (const Vector3 &A, float p) {
 	return Vector3(A.x/p, A.y/p, A.z/p);
 }
 
-double Dot(const Vector3 &A, const Vector3 &B) {
+float Dot(const Vector3 &A, const Vector3 &B) {
 	return A.x*B.x + A.y*B.y + A.z*B.z;
 }
 
 
-double Length(const Vector3 &A) {
+float Length(const Vector3 &A) {
 	return sqrt(Dot(A, A));
 }
 
-double DistanceP2P(const Point3 &A, const Point3 &B) {
+float DistanceP2P(const Point3 &A, const Point3 &B) {
 	return Length(B-A);	
 }
 
-double Angle(const Vector3 &A, const Vector3 &B) {
+float Angle(const Vector3 &A, const Vector3 &B) {
 	return acos(Dot(A, B)/Length(A)/Length(B));
 }
 
-double DistanceToPlane(const Point3 &p, const Point3 &p0, const Vector3 &n) {
+float DistanceToPlane(const Point3 &p, const Point3 &p0, const Vector3 &n) {
 	return fabs(Dot(p-p0, n));
 }
 
@@ -59,7 +59,7 @@ Point3 LinePlaneIntersection(
     const Vector3 &n)
 {
 	Vector3 v = p2 - p1;
-	double t = (Dot(n, p0-p1) / Dot(n, p2-p1)); // denominator = 0?
+	float t = (Dot(n, p0-p1) / Dot(n, p2-p1)); // denominator = 0?
 	return p1 + v*t;
 }
 
@@ -67,19 +67,19 @@ Vector3 Cross(const Vector3 &A, const Vector3 &B) {
 	return Vector3(A.y*B.z - A.z*B.y, A.z*B.x - A.x*B.z, A.x*B.y - A.y*B.x);
 }
 
-double Area2(const Point3 &A, const Point3 &B, const Point3 &C) {
+float Area2(const Point3 &A, const Point3 &B, const Point3 &C) {
 	return Length(Cross(B-A, C-A));
 }
 
 /*
 bool PointInTri(Point3 P, Point3 P0, Point3 P1, Point3 P2) {
-	double area1 = Area2(P, P0, P1);
-	double area2 = Area2(P, P1, P2);
-	double area3 = Area2(P, P2, P0);
+	float area1 = Area2(P, P0, P1);
+	float area2 = Area2(P, P1, P2);
+	float area3 = Area2(P, P2, P0);
 	return dcmp(area1+area2+area3 - Area2(P0, P1, P2)) == 0;
 }
 
-double Volume6(Point3 A, Point3 B, Point3 C, Point3 D) {
+float Volume6(Point3 A, Point3 B, Point3 C, Point3 D) {
 	return Dot(D-A, Cross(B-A, C-A));
 }
 */
@@ -92,10 +92,10 @@ bool isPointInPolyhedron(const Point3 &p, const std::vector<Face> &fs) {
 	for (Face const &f:fs) {
 		Vector3 p2f = f.v[0]-p;
 		Vector3 A = f.normal();
-		double sign = Dot(p2f, f.normal());
+		float sign = Dot(p2f, f.normal());
 		sign /= Length(p2f); // normalization, for stability
 
-		constexpr double bound = -1e-15;
+		constexpr float bound = -1e-15;
 		if(sign < 0.0) return false;
 	}
 	return true;
